* Uncleaned notebook for Video OSD challenge (AI Saturdays Madrid, Week 3).
A CNN classifier based on Resnet-34 was trained using Luke Skywalker and Darth Vader images scraped using Google Images. We then used OpenCV to read frames from the classic Luke/Vader fight scene, fed each frame into the classifier and displayed the two categories alongside their estimated probabilites on top of the frame, creating a new video with the annotated frames.

In our first implementation we simply displayed the name of the most probable category (Vader or Luke), but we then decided to display the exact numbers to better understand the model's performance. The classifier is better at detecting Vader rather than Luke. We believe the dark nature of most frames to be one of the reasons for that, since Vader photos tend to be dark as well.

Some ideas for further work include:

- Improve the quality of the data set. Download more images, and remove the ones that are not representative (illustrations instead of photos, pictures that do not portray either character, etc.).
- Assign a higher probability threshold to Luke so we demand the model to be more confident about that category in order to accept those estimations.
- As David suggested, two networks could be used instead of one, and their results combined: the first network would detect Vader / No Vader; the second would learn to detect frames that contain Luke.

The URLs of the images we scraped have been included with the notebook.
The OpenCV code we used to read and write video files was based on David Griñán's snippet. Thanks a lot!